+++ packages/leches/src/components/GraphControl.vue
@@ -11,7 +11,7 @@ const props = defineProps<{
 
 const containerRef = ref<HTMLElement | null>(null)
 const width = ref(160)
-const height = ref(40)
+const height = ref(20)
 const strokeWidth = 2
 const updateInterval = 100 // Update interval in milliseconds
 const paddingVertical = 4 // Padding at top and bottom to prevent cutting off
@@ -124,7 +124,7 @@ onUnmounted(() => {
 </script>
 
 <template>
-  <div class="tl-flex tl-px-4 tl-gap-1 tl-items-center tl-mb-2">
+  <div class="tl-flex tl-gap-1 tl-items-center" style="padding: 0 var(--tl-h-padding); margin-bottom: var(--tl-unit-spacing);">
     <ControlLabel
       :label="label"
       :control="control"
@@ -135,9 +135,7 @@ onUnmounted(() => {
         tl-relative
         tl-w-2/3
         tl-py-1
-        tl-rounded
         tl-text-right
-        tl-text-xs
         tl-text-gray-400
         tl-bg-gray-100
         dark:tl-bg-dark-300
@@ -146,8 +144,9 @@ onUnmounted(() => {
         tl-border-none
         tl-font-sans
       "
+      style="border-radius: var(--tl-blade-radius);"
     >
-      <div class="tl-absolute tl-bg-gray-100 tl-p-0.5 rounded dark:tl-bg-dark-300 tl-bottom-0.5 tl-right-1 tl-font-sans tl-text-xs">
+      <div class="tl-absolute tl-bg-gray-100 tl-p-0.5 rounded dark:tl-bg-dark-300 tl-bottom-0.5 tl-right-1 tl-font-sans">
         {{ Math.round(control.value) }}
       </div>
       <svg