+++ packages/rapier/src/components/RigidBody.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { ActiveCollisionTypes } from '@dimforge/rapier3d-compat'
-import { type TresObject3D, useLoop } from '@tresjs/core'
+import { useLoop } from '@tresjs/core'
 import { Object3D } from 'three'
 import {
   nextTick,
@@ -13,7 +13,7 @@ import {
 } from 'vue'
 
 import { useRapierContext } from '../composables'
-import { createColliderPropsFromObject, createRigidBody } from '../core'
+import { createRigidBody, createRigidBodyAutoColliderPropsFromObject } from '../core'
 import { hasValidColliderGeometry } from '../utils'
 import { makePropsWatcherRB } from '../utils/props'
 import { Collider } from './colliders'
@@ -41,13 +41,16 @@ const props = withDefaults(defineProps<Partial<RigidBodyProps>>(), {
   enabledTranslations: () => [true, true, true],
 
   // Auto-generated colliders props
+  activeContactForce: false,
+  contactForceEventThreshold: 0,
   friction: 0.5,
   mass: 1,
   restitution: 0,
   density: 1,
   activeCollision: false,
   activeCollisionTypes: ActiveCollisionTypes.DEFAULT,
-  collisionGroups: undefined, // TODO: Make the `collisionGroups` (Not working yet).
+  collisionGroups: undefined,
+  solverGroups: undefined,
   sensor: false,
 })
 
@@ -92,9 +95,10 @@ watch(bodyGroup, async (group) => {
       // Skip children without valid geometry (e.g., collider wrappers, empty Object3Ds)
       if (!hasValidColliderGeometry(child as Object3D)) { continue }
 
-      const createdProps = createColliderPropsFromObject(
-        child as TresObject3D,
+      const createdProps = createRigidBodyAutoColliderPropsFromObject(
+        child,
         props.collider,
+        newPhysicsState.rigidBody,
       )
       collidersProps.push({ ...props, ...createdProps })
     }
@@ -122,7 +126,6 @@ makePropsWatcherRB(props, [
 // reactively set autoColliderProps
 const setAutoColliderProp = <K extends keyof ColliderProps>(prop: K, value: ColliderProps[K]) => {
   if (autoColliderProps.value.length === 0 || props.collider === false) { return }
-
   autoColliderProps.value.forEach((_props) => {
     _props[prop] = value
   })
@@ -136,7 +139,11 @@ watch(() => props.density, value => setAutoColliderProp('density', value))
 watch(() => props.activeCollision, value => setAutoColliderProp('activeCollision', value))
 watch(() => props.activeCollisionTypes, value => setAutoColliderProp('activeCollisionTypes', value))
 watch(() => props.collisionGroups, value => setAutoColliderProp('collisionGroups', value))
+watch(() => props.solverGroups, value => setAutoColliderProp('solverGroups', value))
 watch(() => props.sensor, value => setAutoColliderProp('sensor', value))
+watch(() => props.activeContactForce, value => setAutoColliderProp('activeContactForce', value))
+watch(() => props.contactForceEventThreshold, value => setAutoColliderProp('contactForceEventThreshold', value))
+
 
 watch([() => props.lockTranslations, instance], ([_lockTranslations, _]) => {
   if (!instance.value) { return }
@@ -194,7 +201,10 @@ onUnmounted(() => {
       :activeCollision="_props.activeCollision"
       :activeCollisionTypes="_props.activeCollisionTypes"
       :collisionGroups="_props.collisionGroups"
+      :solverGroups="_props.solverGroups"
       :sensor="_props.sensor"
+      :activeContactForce="_props.activeContactForce"
+      :contactForceEventThreshold="_props.contactForceEventThreshold"
     />
     <slot></slot>
   </TresGroup>