+++ packages/rapier/src/components/colliders/index.ts
@@ -1,26 +1,47 @@
 /** @description `Collider` with the shape set to `ball` */
-export { default as BallCollider } from './BallCollider.vue'
+export { default as BallCollider } from './Ball.vue'
 
 /** @description `Collider` with no shape set */
-export { default as Collider } from './BaseCollider.vue'
+export { default as Collider } from './Base.vue'
 
 /** @description `Collider` with the shape set to `capsule` */
-export { default as CapsuleCollider } from './CapsuleCollider.vue'
+export { default as CapsuleCollider } from './Capsule.vue'
 
 /** @description `Collider` with the shape set to `cone` */
-export { default as ConeCollider } from './ConeCollider.vue'
-
-/** @description `Collider` with the shape set to `cuboid` */
-export { default as CuboidCollider } from './CuboidCollider.vue'
-
-/** @description `Collider` with the shape set to `cylinder` */
-export { default as CylinderCollider } from './CylinderCollider.vue'
-
-/** @description `Collider` with the shape set to `heightfield` */
-export { default as HeightfieldCollider } from './HeightfieldCollider.vue'
+export { default as ConeCollider } from './Cone.vue'
 
 /** @description `Collider` with the shape set to `hull` */
-export { default as HullCollider } from './HullCollider.vue'
+export { default as ConvexHullCollider } from './ConvexHull.vue'
+
+/** @description `Collider` with the shape set to `convexMesh` */
+export { default as ConvexMeshCollider } from './ConvexMesh.vue'
+
+/** @description `Collider` with the shape set to `cuboid` */
+export { default as CuboidCollider } from './Cuboid.vue'
+
+/** @description `Collider` with the shape set to `cylinder` */
+export { default as CylinderCollider } from './Cylinder.vue'
+
+/** @description `Collider` with the shape set to `heightfield` */
+export { default as HeightfieldCollider } from './Heightfield.vue'
+
+/** @description `Collider` with the shape set to `polyline` */
+export { default as PolylineCollider } from './Polyline.vue'
+
+/** @description `Collider` with the shape set to `roundCone` */
+export { default as RoundConeCollider } from './RoundCone.vue'
+
+/** @description `Collider` with the shape set to `roundConeConvexHull` */
+export { default as RoundConvexHullCollider } from './RoundConvexHull.vue'
+
+/** @description `Collider` with the shape set to `roundConeConvexMesh` */
+export { default as RoundConvexMeshCollider } from './RoundConvexMesh.vue'
+
+/** @description `Collider` with the shape set to `roundCuboid` */
+export { default as RoundCuboidCollider } from './RoundCuboid.vue'
+
+/** @description `Collider` with the shape set to `roundCylinder` */
+export { default as RoundCylinderCollider } from './RoundCylinder.vue'
 
 /** @description `Collider` with the shape set to `trimesh` */
-export { default as TrimeshCollider } from './TrimeshCollider.vue'
+export { default as TrimeshCollider } from './Trimesh.vue'